Updated test cases for this crown court outcome.
