+++ maat-court-data-api/src/test/java/gov/uk/courtdata/hearing/crowncourt/validator/CrownCourtValidationProcessorTest.java
@@ -10,15 +10,11 @@ import org.mockito.MockitoAnnotations;
 import org.mockito.junit.MockitoJUnitRunner;
 
 import static gov.uk.courtdata.enums.CrownCourtAppealOutcome.SUCCESSFUL;
-import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 
 @RunWith(MockitoJUnitRunner.class)
 public class CrownCourtValidationProcessorTest {
-
-//    @Mock
-//    private CrownCourtOutComesValidator crownCourtOutComesValidator;
 
     @Mock
     private OUCodeValidator ouCodeValidator;
@@ -33,8 +29,6 @@ public class CrownCourtValidationProcessorTest {
     }
 
 
-
-
     @Test
     public void givenHearingData_whenCrownCourtValidationProcessorIsCalled_thenValidatorsAreCalled() {
 
@@ -45,11 +39,7 @@ public class CrownCourtValidationProcessorTest {
         // when
         crownCourtValidationProcessor.validate(hearingDetails, SUCCESSFUL.getValue());
 
-        //then
-        //verify(crownCourtOutComesValidator, times(1)).validate(hearingDetails);
-
         verify(ouCodeValidator, times(1)).validate(hearingDetails);
-        verify(caseTypeValidator, times(1)).validate(hearingDetails,anyString());
-
+        verify(caseTypeValidator, times(1)).validate(hearingDetails,SUCCESSFUL.getValue());
     }
 }
